+++ exonum-java-binding-core/src/test/java/com/exonum/binding/transport/VertxServerIntegrationTest.java
@@ -27,6 +27,8 @@ import io.vertx.core.buffer.Buffer;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.client.HttpResponse;
 import io.vertx.ext.web.client.WebClient;
+import java.io.IOException;
+import java.net.ServerSocket;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
@@ -41,7 +43,7 @@ public class VertxServerIntegrationTest {
   @Rule
   public ExpectedException expectedException = ExpectedException.none();
 
-  private static final int PORT = 0;
+  private static final int ANY_PORT = 0;
 
   private VertxServer server;
 
@@ -60,7 +62,7 @@ public class VertxServerIntegrationTest {
 
   @Test
   public void mountSubRouter_stoppedServer() throws Exception {
-    server.start(PORT);
+    server.start(ANY_PORT);
     Router router = server.createRouter();
     blockingStop();
 
@@ -71,10 +73,10 @@ public class VertxServerIntegrationTest {
   @Test
   public void start_WontStartTwice() throws Exception {
     try {
-      server.start(PORT);
+      server.start(ANY_PORT);
 
       expectedException.expect(IllegalStateException.class);
-      server.start(PORT);
+      server.start(ANY_PORT);
     } finally {
       blockingStop();
     }
@@ -82,7 +84,7 @@ public class VertxServerIntegrationTest {
 
   @Test
   public void stop_properlyStops() throws Exception {
-    server.start(PORT);
+    server.start(ANY_PORT);
     CompletableFuture<Void> f = server.stop();
     f.get(4, TimeUnit.SECONDS);
     assertTrue(f.isDone());
@@ -90,7 +92,7 @@ public class VertxServerIntegrationTest {
 
   @Test
   public void stop_subsequentStopsHaveNoEffect() throws Exception {
-    server.start(PORT);
+    server.start(ANY_PORT);
     CompletableFuture<Void> f = server.stop();
     f.get(5, TimeUnit.SECONDS);
 
@@ -100,11 +102,11 @@ public class VertxServerIntegrationTest {
 
   @Test
   public void start_wontStartStopped() throws Exception {
-    server.start(PORT);
+    server.start(ANY_PORT);
     blockingStop();
 
     expectedException.expect(IllegalStateException.class);
-    server.start(PORT);
+    server.start(ANY_PORT);
   }
 
   @Test
@@ -112,7 +114,7 @@ public class VertxServerIntegrationTest {
     Vertx wcVertx = null;
     try {
       // Start a server.
-      int port = 8080;
+      int port = findFreePort();
       server.start(port);
 
       // Define a request handler.
@@ -164,4 +166,14 @@ public class VertxServerIntegrationTest {
     int stopTimeout = 2;
     f.get(stopTimeout, TimeUnit.SECONDS);
   }
+
+  /**
+   * Returns random available local port.
+   */
+  private int findFreePort() throws IOException {
+    try (ServerSocket socket = new ServerSocket(ANY_PORT)) {
+      return socket.getLocalPort();
+    }
+  }
+
 }